Clean up push notification logic
